Should fix some gcc 64bit issues.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2450 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/unzip.c
@@ -142,7 +142,7 @@ local int unzlocal_getShort(vfsfile_t *fin,unsigned long *pi) {
 }
 
 local int unzlocal_getLong(vfsfile_t *fin,unsigned long *pi) {
-	unsigned long c;
+	unsigned int c;
 	int err = VFS_READ(fin, &c, 4);
     if (err==4) {
 	    *pi = LittleLong(c);
